+++ src/app/components/cards/cart-table-sentiments/cart-table-sentiments.component.html
@@ -3,34 +3,7 @@
 <div class="relative grid grid-flow-row gap-4 md:pt-16 pb-16">
 
 <h2 class=" text-2xl font-bold">Sentimental Analytic </h2>
-<div class="container" #container>
-	<div class="left fa-solid fa-arrow-left" *ngIf="totalPages > 1" (click)="currentPage != 1 && changePage(-1)" [ngClass]="{'disabled':currentPage === 1}">
- </div>
-  <div class="cards" [ngClass]="{'w-100': totalPages === 1}">
-		<div class="overflow"  [ngStyle]="{'width': overflowWidth, 'left': pagePosition}">
-			<div class=" card    "   *ngFor="let ele of data "   [ngStyle]="{'width': cardWidth , 'background-color':getColor(ele)}">
-  
-        <div class="flex items-start "> 
 
-        
-          <div class="relative w-auto my-4 pl-4 flex-initial">
-            <div class="p-3 text-center inline-flex items-center justify-center w-12 h-12 text-2xl text-blueGray-400">
-              <i [ngClass]="getIcon(ele)"></i>
-            </div>
-          </div>
-    
-        <div class=" flex flex-col items-start justify-between ml-4 p-4">
-          <h2 class="font-semibold">{{  (((ele.split(',')[0] ) * 100 ) / allComments ).toFixed(0)}} %</h2>
-          <p class=" text-md text-black"> {{ele.split(',')[1]}} Comment Rate</p>
-        </div>
-      </div>
-    
-      </div>
-		</div>
-	</div>
-	<div class="right  fa-solid fa-arrow-right" *ngIf="totalPages > 1" (click)="currentPage != totalPages && changePage(+1)"
-		[ngClass]="{'disabled':currentPage === totalPages}"></div>
-</div>
 
 
 <div  class="mt-4 grid grid-flow-row grid-cols-4 gap-4  sm:grid-row-2 lg:grid-row-3 xl:grid-row-4">
