+++ src/app/components/cards/cart-table-topic/cart-table-topic.component.html
@@ -1,0 +1,163 @@
+<div class="relative grid grid-flow-row gap-4 md:pt-16 pb-16">
+
+  <h2 class=" text-2xl font-bold">Topic Analytic </h2>
+
+  
+
+
+  <div class="mt-4 grid grid-flow-row grid-cols-4 gap-4  sm:grid-row-2 lg:grid-row-3 xl:grid-row-4">
+
+
+
+
+
+  </div>
+
+
+  <div style="height: 73vh ;"
+    class="relative overflow-x-auto flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded"
+    [ngClass]="color === 'light' ? 'bg-white' : 'bg-red-900 text-white'">
+    <div class="rounded-t   w-fit  mb-0 px-4 py-3 border-0">
+      <div class="flex flex-wrap items-center justify-center">
+        <div class="relative px-4 max-w-full flex flex-grow flex-row justify-between  items-center">
+          <h3 class="font-semibold text-lg p-4 mb-4" [ngClass]="color === 'light' ? 'text-blueGray-700' : 'text-white'">
+            Comments By Emotions
+          </h3>
+
+
+
+
+        </div>
+      </div>
+      <div *ngIf="!Comments">
+
+        <div role="status"
+          class="p-4 space-y-4  max-w-3xl rounded border border-gray-200 divide-y divide-gray-200 shadow animate-pulse dark:divide-gray-700 md:p-6 dark:border-gray-700">
+          <div class="flex justify-between items-center">
+            <div>
+              <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-600 w-24 mb-2.5"></div>
+              <div class="w-32 h-2 bg-gray-200 rounded-full dark:bg-gray-700"></div>
+            </div>
+            <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-700 w-12"></div>
+          </div>
+          <div class="flex justify-between items-center pt-4">
+            <div>
+              <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-600 w-24 mb-2.5"></div>
+              <div class="w-32 h-2 bg-gray-200 rounded-full dark:bg-gray-700"></div>
+            </div>
+            <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-700 w-12"></div>
+          </div>
+          <div class="flex justify-between items-center pt-4">
+            <div>
+              <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-600 w-24 mb-2.5"></div>
+              <div class="w-32 h-2 bg-gray-200 rounded-full dark:bg-gray-700"></div>
+            </div>
+            <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-700 w-12"></div>
+          </div>
+          <div class="flex justify-between items-center pt-4">
+            <div>
+              <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-600 w-24 mb-2.5"></div>
+              <div class="w-32 h-2 bg-gray-200 rounded-full dark:bg-gray-700"></div>
+            </div>
+            <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-700 w-12"></div>
+          </div>
+          <div class="flex justify-between items-center pt-4">
+            <div>
+              <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-600 w-24 mb-2.5"></div>
+              <div class="w-32 h-2 bg-gray-200 rounded-full dark:bg-gray-700"></div>
+            </div>
+            <div class="h-2.5 bg-gray-300 rounded-full dark:bg-gray-700 w-12"></div>
+          </div>
+          <span class="sr-only">Loading...</span>
+        </div>
+
+      </div>
+
+      <!-- Projects table -->
+      <table datatable [dtOptions]="dtOptions" [dtTrigger]="dtTrigger"
+        class="items-center    bg-transparent border-collapse">
+        <thead *ngIf="Comments">
+          <tr>
+            <th
+              class="px-6 align-middle  border border-solid py-3 text-xs uppercase border-l-0 border-r-0  font-semibold text-left"
+              [ngClass]="
+                  color === 'light'
+                    ? 'bg-blueGray-50 text-blueGray-500 border-blueGray-100'
+                    : 'bg-red-800 text-red-300 border-red-700'
+                ">
+              Date
+            </th>
+            <th
+              class="px-6 align-middle  border border-solid py-3 text-xs uppercase border-l-0 border-r-0  font-semibold text-left"
+              [ngClass]="
+                  color === 'light'
+                    ? 'bg-blueGray-50 text-blueGray-500 border-blueGray-100'
+                    : 'bg-red-800 text-red-300 border-red-700'
+                ">
+              Commentaire
+            </th>
+            <th
+              class="px-6 align-middle border border-solid py-3 text-xs uppercase border-l-0 border-r-0  font-semibold text-left"
+              [ngClass]="
+                  color === 'light'
+                    ? 'bg-blueGray-50 text-blueGray-500 border-blueGray-100'
+                    : 'bg-red-800 text-red-300 border-red-700'
+                ">
+              Source
+            </th>
+
+            <th
+              class="px-6 align-middle border border-solid py-3 text-xs uppercase border-l-0 border-r-0  font-semibold text-left"
+              [ngClass]="
+                  color === 'light'
+                    ? 'bg-blueGray-50 text-blueGray-500 border-blueGray-100'
+                    : 'bg-red-800 text-red-300 border-red-700'
+                ">
+              Topic
+            </th>
+
+
+          </tr>
+        </thead>
+        <tbody id="excel-table">
+          <tr *ngFor="let user of Comments" class="border-b">
+            <th class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs p-4 text-left flex items-center">
+
+              <span class="ml-3 font-bold" [ngClass]="color === 'light' ? 'text-blueGray-600' : 'text-white'">
+                {{ user.date.split('T')[0] }}
+              </span>
+            </th>
+
+
+            <td class="border-t-0 break-word  px-6 align-middle  border-l-0 border-r-0 text-xs  p-4">
+              {{ user.commentaire }}
+
+            </td>
+            <td class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs  p-4">
+              {{ user.source | uppercase }}
+
+            </td>
+
+            <td class="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs  p-4">
+              <div class="flex items-center">
+                <span class="mr-2">{{user.topic | uppercase}}</span>
+
+              </div>
+            </td>
+
+
+          </tr>
+
+
+        </tbody>
+      </table>
+
+
+
+
+
+
+    </div>
+
+  </div>
+</div>
